Add note on tool click

+++ S2VX.Game/Story/Approach.cs
@@ -9,7 +9,7 @@ namespace S2VX.Game.Story
 {
     public class Approach : CompositeDrawable
     {
-        public float EndTime { get; set; } = 0;
+        public double EndTime { get; set; } = 0;
         public Vector2 Coordinates { get; set; } = Vector2.Zero;
 
         [Resolved]
